Move DataSegregationStrategy type to metabase-types/api/permissions
This is a permissions concept, not an embedding one. Moving it to
metabase-types resolves a module boundary violation (basic/api importing
from shared/other) and breaks a circular dependency between metabase/api
and metabase/embedding/embedding-hub.
