+++ enterprise/frontend/src/metabase-enterprise/tenants/components/TenantsSummaryOnboardingStep/TenantsSummaryOnboardingStep.tsx
@@ -4,11 +4,11 @@ import { push } from "react-router-redux";
 import { jt, msgid, ngettext, t } from "ttag";
 
 import { RelatedSettingCard } from "metabase/admin/components/RelatedSettingsSection";
-import type { DataSegregationStrategy } from "metabase/embedding/embedding-hub";
 import { conjunct } from "metabase/lib/formatting/strings";
 import { useDispatch } from "metabase/lib/redux";
 import type { CreatedTenantData } from "metabase/plugins/oss/tenants";
 import { Button, Flex, SimpleGrid, Stack, Text, Title } from "metabase/ui";
+import type { DataSegregationStrategy } from "metabase-types/api";
 
 import { useListTenantsQuery } from "../../../api/tenants";
 import { getIsolationFieldConfig } from "../CreateTenantsOnboardingStep/isolation-field-config";
