+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/DataSegregationStrategyPicker.tsx
@@ -4,12 +4,9 @@ import { t } from "ttag";
 
 import type { IconName } from "metabase/ui";
 import { Button, Flex, Group, Icon, Radio, Stack, Text } from "metabase/ui";
-
-import type { DataSegregationStrategy } from "../../types";
+import type { DataSegregationStrategy } from "metabase-types/api";
 
 import S from "./DataSegregationStrategyPicker.module.css";
-
-export type { DataSegregationStrategy };
 
 interface DataSegregationStrategyPickerProps {
   value: DataSegregationStrategy | null;
