+++ frontend/src/metabase/embedding/embedding-hub/types/embedding-checklist.ts
@@ -1,10 +1,5 @@
 import type { AddDataTab } from "metabase/nav/containers/MainNavbar/MainNavbarContainer/AddDataModal/utils";
 
-export type DataSegregationStrategy =
-  | "row-column-level-security"
-  | "connection-impersonation"
-  | "database-routing";
-
 export type EmbeddingHubStepId =
   | "create-test-embed"
   | "add-data"
