+++ frontend/src/metabase/embedding/embedding-hub/types/index.ts
@@ -1,5 +1,4 @@
 export type {
-  DataSegregationStrategy,
   EmbeddingHubStep,
   EmbeddingHubStepId,
   EmbeddingHubAction,
